+++ app/src/main/java/com/example/ourcourage/android/presentation/ui/home/HomeScreen.kt
@@ -12,7 +12,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.ourcourage.android.R
-import com.example.ourcourage.android.data.MultiUseList.multiUseList
+import com.example.ourcourage.android.data.datasource.local.MultiUseList.multiUseList
 import com.example.ourcourage.android.presentation.ui.component.OurCourageTopBarText
 import com.example.ourcourage.android.presentation.ui.home.list.HomeMultiUseHistoryList
 import com.example.ourcourage.android.ui.theme.BackgroundBlue
