+++ app/src/main/java/com/example/ourcourage/android/presentation/ui/point/PointHistoryScreen.kt
@@ -22,7 +22,7 @@ import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.example.ourcourage.android.R
-import com.example.ourcourage.android.data.MultiUseList
+import com.example.ourcourage.android.data.datasource.local.MultiUseList
 import com.example.ourcourage.android.domain.MultiUse
 import com.example.ourcourage.android.presentation.ui.component.OurCourageTopBarText
 import com.example.ourcourage.android.ui.theme.OurCourageAndroidv2Theme
